requirements.txt
asgiref==3.5.2
django-extensions==3.1.5
Django==4.0.4
Pillow==9.1.1
pip==22.1.1
psycopg2==2.9.3
setuptools==60.2.0
sqlparse==0.4.2
tzdata==2022.1
wheel==0.37.1

djangorestframework~=3.13.1